+++ packages/seacas/libraries/ioss/src/exodusII/Ioex_SuperElement.C
@@ -166,7 +166,7 @@ Ioex::SuperElement::SuperElement(const std::string &filename,
   // but for now we only need "Kr" and "Mr"
 }
 
-int Ioex::SuperElement::internal_get_field_data(const Ioss::Field& field,
+int64_t Ioex::SuperElement::internal_get_field_data(const Ioss::Field& field,
 				      void *data, size_t data_size) const
 {
   size_t num_to_get = field.verify(data_size);
@@ -203,7 +203,7 @@ Ioex::SuperElement::~SuperElement()
   if (filePtr) nc_close(filePtr);
 }
 
-int Ioex::SuperElement::internal_put_field_data(const Ioss::Field& /* field */,
+int64_t Ioex::SuperElement::internal_put_field_data(const Ioss::Field& /* field */,
 						void* /* data */, size_t /* data_size */) const
 {
   return -1;
